requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
multidict==6.0.5
openai==1.16.2
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
python-weather==2.0.1
sniffio==1.3.1
termcolor==2.4.0
tqdm==4.66.2
typing_extensions==4.10.0
yarl==1.9.4
